+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## main
 
 - Added `Tello.Manger.client_pids/1` to fetch the running `Tello.Client` PIDs.
+- Use `integer` instead of `reference` as `Tello.Client` and it's children's name.
 
 ## v0.2.0